requirements.txt
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.54.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.4
gspread-dataframe==4.0.0
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
WooCommerce==3.0.0
